initial commit
The kernel is basically complete, except that a few more filesystem,
path primitives, and environment-variable primitives are needed.
Otherwise, enough is here to run

 #lang zuo/private/defmacromod

 (letrec ([fib (lambda (n)
                 (if (= n 0)
                     1
                     (if (= n 1)
                         1
                         (+ (fib (- n 1)) (fib (- n 2))))))])
   (display (~a (fib 30) "\n")))

Thee next step is to fill in the module and macro part of
`defmacrmod`. The step after that is to fill out a standard library.
And the step after that is to create a `make` DSL, which is whole
point.