split into two languages: hygienic macros and not
Hygienic macros are better, of course, but scope-set machinery on a
slow interpreter is too slow (so far, at least). It's not difficult to
have both system, because there's one expander implementation
parameterized over the representation of syntax. The base libraries so
far are similarly parameterized. In both cases, the parameterization
boils down to S-expression reuse via `include`.

Currently, the default `#lang zuo` language is fast and non-hygienic,
and I suspect that will be the useful one. The alternative
implementation is `#lang zuo/hygiene`, which is maybe useful as a demo
implementation of scope sets that runs on a very small base.

Even the non-hygienic macro system has syntax-object identifiers as
syntactic closures. (Since there are only top-level macros, the
closure just identifies the enclosing module.) So, by being careful
enough, we can write useful macros in that system.