move away from absolute paths for modules
Since the current directory can never changed, a relative path can
identify a module just as well as an absolute path. We just have to
take care to perform the same normalization of "." and ".." as before
in `build-module-path`.

Meanwhile, in `zuo/build`, use absolute paths in "_zuo_ts.db" when an
imput path is specified by an absolute path. This avoids problems with
trying to turn a "/tmp" path into a relative path on Mac OS, for
example, where "/tmp" is link to "/private/tmp" (and so syntactic path
normalizations don't line up with the filesystem's resolution of a
path).