yet another try at path manipulation
When a path includes "..", there's a difference between syntactically
simplifying a path, such as converting "x/../y" to just "y", and
paying attention to the file system, where "x" might be a link to
"z/w" so that "x/../y" accesses the same file "z/y". In a real sense,
syntactic resolution just doesn't work.

Nevertheless, we want to syntactically manipulate paths all the time.
Modules need to have consistent names, and paths need to be converted
to and from relative to some other path for recording build results.

This revision doubles down on syntactic manipulation by making
`build-path` normalize added elements. Normalization removes "." and
resolves ".." in path additions (and it also normalizing path
separators). It leaves the base path alone, however, except as needed
to avoid adding ".."s to the end. So, if you use a path with links as
a starting point and path additions stay within the subtree that is is
the target of a link, then things should work right.

The `build-raw-path` function provides non-normalizing functionality,
just in case that's useful.