adjust CI tests, repair more warnings, fix `fd-poll` for Windows
Change the CI to check both normal compilation and compialtion with
`-DZUO_EMBEDDED`. Use Ubuntu 22.04 for CI, since its compiler seems to
be more picky. Add `-D_POSIX_C_SOURCE=200809L` with `-std=c11`,
because that combination is needed on some platforms (including the
one used for CI). Fix macOS compilation when `_POSIX_C_SOURCE` is
defined.

Repair a bug in Windows for `fd-poll`, but also update test and docs
to clarify that `fd-poll` is not so useful on Windows. Add Windows
to CI.